Initial refactoring completed
