+++ requirements/9.md
@@ -12,7 +12,7 @@
   - `Hyeri Heo`
 
 ## Team Requirements
-The main goal is to display essential in-game elements on the gameplay screen.
+The main goal is to display essential in-game elements on the gameplay ui.
 
 - Combo
 - Playtime
