+++ src/engine/core/GameSettings.java
@@ -1,4 +1,4 @@
-package engine;
+package engine.core;
 
 /**
  * Implements an object that stores a single game's difficulty settings.
