+++ src/engine/manager/InputManager.java
@@ -1,4 +1,4 @@
-package engine;
+package engine.manager;
 
 import java.awt.event.KeyEvent;
 import java.awt.event.KeyListener;
