+++ src/engine/utility/MinimalFormatter.java
@@ -1,4 +1,4 @@
-package engine;
+package engine.utility;
 
 import java.text.DateFormat;
 import java.text.SimpleDateFormat;
