+++ src/entity/Entity.java
@@ -2,7 +2,7 @@ package entity;
 
 import java.awt.Color;
 
-import engine.DrawManager.SpriteType;
+import engine.manager.DrawManager.SpriteType;
 
 /**
  * Implements a generic game entity.
