+++ src/entity/ItemBox.java
@@ -1,8 +1,8 @@
 package entity;
 
-import engine.DrawManager.SpriteType;
-import engine.Sound;
-import engine.SoundManager;
+import engine.manager.DrawManager.SpriteType;
+import engine.utility.Sound;
+import engine.manager.SoundManager;
 
 import java.awt.*;
