+++ src/entity/Ship.java
@@ -3,11 +3,11 @@ package entity;
 import java.awt.Color;
 import java.util.Set;
 
-import engine.Cooldown;
-import engine.Core;
-import engine.DrawManager.SpriteType;
-import engine.Sound;
-import engine.SoundManager;
+import engine.utility.Cooldown;
+import engine.core.Core;
+import engine.manager.DrawManager.SpriteType;
+import engine.utility.Sound;
+import engine.manager.SoundManager;
 
 /**
  * Implements a ship, to be controlled by the player.
