+++ src/ui/CreditScreen.java
@@ -1,12 +1,12 @@
-package screen;
+package ui;
 
 import java.awt.event.KeyEvent;
 import java.io.IOException;
 import java.util.List;
 
-import engine.Core;
-import engine.Sound;
-import engine.SoundManager;
+import engine.core.Core;
+import engine.utility.Sound;
+import engine.manager.SoundManager;
 
 public class CreditScreen extends Screen{
