+++ src/ui/ShopScreen.java
@@ -1,11 +1,11 @@
-package screen;
+package ui;
 
 import java.awt.event.KeyEvent;
 
-import engine.Cooldown;
-import engine.Core;
-import engine.Sound;
-import engine.SoundManager;
+import engine.utility.Cooldown;
+import engine.core.Core;
+import engine.utility.Sound;
+import engine.manager.SoundManager;
 import entity.Wallet;
 
 public class ShopScreen extends Screen {
