+++ src/ui/WelcomeFrame.java
@@ -1,4 +1,4 @@
-package engine;
+package ui;
 
 import javax.swing.*;
 import java.awt.*;
